issue-95: fixed mistake in money percent increase calculations

+++ SpaceInvaders/src/main/java/com/gmail/yaroslavlancelot/spaceinvaders/gameobjects/objects/staticobjects/PlanetStaticObject.java
@@ -55,7 +55,7 @@ public class PlanetStaticObject extends StaticObject {
             }
             value += building.getIncome();
         }
-        return value + value * (percentIncrease / 100);
+        return value + (int) (value * (((float)percentIncrease) / 100));
     }
 
     /** set unit spawn point */
